### 2025-08-07 #8762 @crvernon
:wave: @wbenoit26 - thank you for your submission to JOSS.  We are working through a large queue of submissions right now and will get a topic editor assigned for you as soon as one becomes available.  Thank you.
