### 2025-08-18 46ee023
Added Alec as author and included additional section headings
