### 2025-08-18 af8f15c
Tweak working and correct PyTorch citation
