+++ joss_submission/references.bib
@@ -34,7 +34,7 @@ archivePrefix = {arXiv},
        doi            = "10.7935/GT1W-FZ16",
        year           = "2018"
  }
-@misc{paszke2019pytorch,
+@inproceedings{paszke2019pytorch,
   title={PyTorch: An Imperative Style, High-Performance Deep Learning Library},
   author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalya and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
   booktitle={Advances in Neural Information Processing Systems 32},
