### 2025-08-19 #8762 @jatkinson1000
Thanks @wbenoit26

The decision is left to the authors, but like with any work JOSS believes credit should be given to anyone who contributed towards it in a significant manner. Thanks for this.

You can regenerate the version of the paper using the [editorial-bot commands](https://joss.readthedocs.io/en/latest/editorial_bot.html#the-complete-list-of-available-commands).
I'll do this for you now, but you can do so whenever you update in future.
