### 2025-08-20 #8836 @editorialbot
[REVIEW]: ml4gw: PyTorch utilities for training neural networks in gravitational wave physics applications
**Submitting author:** @wbenoit26 (<a href="http://orcid.org/0000-0003-4750-9413">William Benoit</a>)
**Repository:** https://github.com/ML4GW/ml4gw
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v0.7.8
