### 2025-08-20 ae8fc85
Use more accurate knot placement in spline interpolation
