### 2025-08-20 f872e43
Remove old testing comments
