### 2025-08-21 d07a3b0
Split interpolate into 1D and 2D, got Q-transforms working
