+++ ml4gw/transforms/spline_interpolation.py
@@ -5,111 +5,14 @@ Adaptation of code from https://github.com/dottormale/Qtransform_torch/
 from typing import Optional, Tuple
 
 import torch
+import torch.nn.functional as F
 from torch import Tensor
-from jaxtyping import Float
 
 
-class SplineInterpolate(torch.nn.Module):
+class SplineInterpolateBase(torch.nn.Module):
     """
-    Perform 1D or 2D spline interpolation based on De Boor's method.
-    Supports batched, multi-channel inputs, so acceptable data
-    shapes are ``(width)``, ``(height, width)``, ``(batch, width)``,
-    ``(batch, height, width)``, ``(batch, channel, width)``, and
-    ``(batch, channel, height, width)``.
-
-    During initialization of this Module, both the desired input
-    and output coordinate Tensors can be specified to allow
-    pre-computation of the B-spline basis matrices, though the only
-    mandatory argument is the coordinates of the data along the
-    ``width`` dimension. If no argument is given for coordinates along
-    the ``height`` dimension, it is assumed that 1D interpolation is
-    desired.
-
-    Unlike scipy's implementation of spline interpolation, the data
-    to be interpolated is not passed until actually calling the
-    object. This is useful for cases where the input and output
-    coordinates are known in advance, but the data is not, so that
-    the interpolator can be set up ahead of time.
-
-    WARNING: compared to scipy's spline interpolation, this method
-    produces edge artifacts when the output coordinates are near
-    the boundaries of the input coordinates. Therefore, it is
-    recommended to interpolate only to coordinates that are well
-    within the input coordinate range. Unfortunately, the specific
-    definition of "well within" changes based on the size of the
-    data, so some testing may be required to get good results.
-
-    Args:
-        x_in:
-            Coordinates of the width dimension of the data
-        y_in:
-            Coordinates of the height dimension of the data. If not
-            specified, it is assumed the 1D interpolation is desired,
-            and so the default value is a Tensor of length 1
-        kx:
-            Degree of spline interpolation along the width dimension.
-            Default is cubic.
-        ky:
-            Degree of spline interpolation along the height dimension.
-            Default is cubic.
-        sx:
-            Regularization factor to avoid singularities during matrix
-            inversion for interpolation along the width dimension. Not
-            to be confused with the ``s`` parameter in scipy's spline
-            methods, which controls the number of knots.
-        sy:
-            Regularization factor to avoid singularities during matrix
-            inversion for interpolation along the height dimension.
-        x_out:
-            Coordinates for the data to be interpolated to along the
-            width dimension. If not specified during initialization,
-            this must be specified during the object call.
-        y_out:
-            Coordinates for the data to be interpolated to along the
-            height dimension. If not specified during initialization,
-            this must be specified during the object call.
-
+    Base class for spline interpolation.
     """
-
-    def __init__(
-        self,
-        x_in: Float[Tensor, " width"],
-        y_in: Float[Tensor, " height"] = None,
-        kx: int = 3,
-        ky: int = 3,
-        sx: float = 0.0,
-        sy: float = 0.0,
-        x_out: Optional[Float[Tensor, " width"]] = None,
-        y_out: Optional[Float[Tensor, " height"]] = None,
-    ):
-        super().__init__()
-        if y_in is None:
-            y_in = Tensor([1])
-        self.kx = kx
-        self.ky = ky
[…]
