### 2025-08-21 8571e91
Re-organize spline tests
