### 2025-08-21 8c264e6
Handle output coordinates that are outside input space
