### 2025-08-21 be45440
Add checks for input coordinate length
