### 2025-08-22 516409b
Enable dynamic versioning
