+++ _sources/api/ml4gw.transforms.spline_interpolation.rst.txt
@@ -8,5 +8,7 @@ ml4gw.transforms.spline\_interpolation
 
    .. autosummary::
    
-      SplineInterpolate
+      SplineInterpolate1D
+      SplineInterpolate2D
+      SplineInterpolateBase
    
