### 2025-08-22 7bfb8c0
drop support for python3.9 (EOL in Oct 2025)
- update workflows
- ensure PEP compliant license field
