+++ .github/workflows/pre-commit.yaml
@@ -13,5 +13,5 @@ jobs:
     - uses: actions/checkout@v4
     - uses: actions/setup-python@v5
       with:
-        python-version: "3.12"
+        python-version: "3.10"
     - uses: pre-commit/action@v3.0.1
