### 2025-08-22 068250d
skip strict zip in ruff
