### 2025-08-22 8d92197
ignore zip check in tool.ruff
