### 2025-08-27 66d0f1e
apply strict zip formatting
