+++ tests/test_gw.py
@@ -67,7 +67,9 @@ def bilby_get_ifo_response(ifos, batch_size):
 
     def func(ra, dec, psi, geocent_time, modes):
         responses = np.zeros((batch_size, len(modes), len(ifos)))
-        for i, (r, d, p, t) in enumerate(zip(ra, dec, psi, geocent_time)):
+        for i, (r, d, p, t) in enumerate(
+            zip(ra, dec, psi, geocent_time, strict=False)
+        ):
             for j, ifo in enumerate(ifos):
                 for k, mode in enumerate(modes):
                     responses[i, k, j] = ifo.antenna_response(r, d, t, p, mode)
@@ -154,9 +156,11 @@ def bilby_shift_responses(
     def func(ra, dec, geocent_time, responses):
         output = np.zeros_like(responses)
         for i, (r, d, t, sample) in enumerate(
-            zip(ra, dec, geocent_time, responses)
+            zip(ra, dec, geocent_time, responses, strict=False)
         ):
-            for j, (response, ifo) in enumerate(zip(sample, ifos)):
+            for j, (response, ifo) in enumerate(
+                zip(sample, ifos, strict=False)
+            ):
                 shifted = do_shift(ifo, r, d, t, response)
                 output[i, j] = shifted
         return output
