+++ ml4gw/dataloading/hdf5_dataset.py
@@ -153,7 +153,9 @@ class Hdf5TimeSeriesDataset(torch.utils.data.IterableDataset):
         unique_fnames, inv, counts = np.unique(
             fnames, return_inverse=True, return_counts=True
         )
-        for i, (fname, count) in enumerate(zip(unique_fnames, counts)):
+        for i, (fname, count) in enumerate(
+            zip(unique_fnames, counts, strict=False)
+        ):
             size = self.sizes[fname]
             max_idx = size - self.kernel_size
 
@@ -185,7 +187,9 @@ class Hdf5TimeSeriesDataset(torch.utils.data.IterableDataset):
             # open the file and sample a different set of
             # kernels for each batch element it occupies
             with h5py.File(fname, "r") as f:
-                for b, c, i in zip(batch_indices, channel_indices, idx):
+                for b, c, i in zip(
+                    batch_indices, channel_indices, idx, strict=False
+                ):
                     x[b, c] = f[self.channels[c]][i : i + self.kernel_size]
         return torch.Tensor(x)
 
