### 2025-09-05 f580ae6
adjust strictness in zip
