### 2025-09-05 8eaeee4
drop support for python3.9 (EOL in Oct 2025) (#222)
* drop support for python3.9 (EOL in Oct 2025)

- update workflows
- ensure PEP compliant license field
- apply strictness in zip functions
