### 2025-09-08 62b7fec
Dataloading and waveform generation buff
