### 2025-09-12 8453c01
added phenom_deco.py, update __init__.py, added phenom_deco.ipynb in examples
