### 2025-09-15 4509c49
update readme and add contribution guide
