### 2025-09-16 #8836 @Uddiptaatwork
### Quick summary:
I am almost done with my review.
- Installation and setup was smooth and runs well
- The paper seems concise and well explained
- The code has it's core functionality described well in the documentation with relevant examples.
---

### One quick thing, and although this is NOT publication blocking, it is a suggestion that I believe enhances the code:
- A `unittest`/`pytest` based testing framework in line with the categories set out in the example usage section of the documentation that highlights a clean return would be appreciated.
- I would also add to the readme of the repo, a couple of lines stating a very small minimal working example along the lines of the usage examples.
---
Other than that, I am happy to accept the article as well as the repo.
