### 2025-09-18 21e5bef
Changed whitening to be able to take different highpass and lowpass params on the forward call
