### 2025-09-18 6074872
Incorporate feedback from JOSS reviewers
