+++ docs/index.rst
@@ -8,14 +8,15 @@ Welcome to ml4gw's documentation!
 
 .. note:: 
    This documentation is a work in progress!
-   If you have any questions or suggestions, please feel free to reach out to ml4gw@ligo.mit.edu
+   If you have any questions or suggestions, please feel free to reach out create an `issue <https://github.com/ML4GW/ml4gw/issues>\_
 
 ml4gw
 =====
 ml4gw is a library of `pytorch  <https://pytorch.org/docs/stable/index.html>`_ utilities 
 for training neural networks in service of gravitational wave physics applications.
 
-The code can be found on github at `<https://github.com/ML4GW/ml4gw>`_
+The code can be found on github at `<https://github.com/ML4GW/ml4gw>`_. Please see the 
+usage examples and tutorial below for help on getting started.
 
 
 .. toctree::
@@ -52,7 +53,6 @@ we anticipate that novel use cases will encounter errors stemming from a lack of
 We encourage users who encounter these difficulties to file issues on GitHub, and we'll be happy to offer 
 support to extend our coverage to new or improved functionality. 
 We also strongly encourage ML users in the GW physics space to try their hand at working on these issues and joining on as collaborators! 
-For more information about how to get involved, feel free to reach out to ml4gw@ligo.mit.edu. 
 By bringing in new users with new use cases, we hope to develop this library into a truly general-purpose tool which makes 
 DL more accessible for gravitational wave physicists everywhere.
    
