### 2025-09-18 c841788
Minor fixes
