### 2025-09-24 76e00e5
Add decimator transform for downsampling timeseries data
Implements a decimation transform to reduce the sampling rate of input time series based on a schedule. Also includes unit test in .
