+++ ml4gw/transforms/decimator.py
@@ -1,79 +1,195 @@
 import torch
 
+
 class Decimator(torch.nn.Module):
     """
     Downsample/decimate timeseries according to a schedule.
 
-    The schedule is a tensor of shape (num_segments, 3), where each row is:
+    The schedule is a tensor of shape (N, 3), where each row is:
     [start_time, end_time, target_sample_rate]
 
     Example default schedule (seconds, Hz):
-    [[0, 40, 256],     # first 40s at 256 Hz
-     [40, 58, 512],    # next 18s at 512 Hz
-     [58, 60, 2048]]   # last 2s at full 2048 Hz
+    [[0, 40, 256],   # first 40s at 256 Hz
+     [40, 58, 512],  # next 18s at 512 Hz
+     [58, 60, 2048]] # last 2s at full 2048 Hz
+
+    Strain: Tensor of shape (B, C, T), where
+    B=batch size, C=channels, T=time.
 
     Functions:
-    build_variable_indices(sr, schedule, device):
-    Compute the sample indices to keep from the full signal.
-
-    split_by_schedule(signal, schedule, device):
-    Split a signal into segments according to the schedule.
+        build_variable_indices(sample_rate, schedule, device)
+            Compute the sample indices to keep from the full strain.
+        split_by_schedule(strain, schedule)
+            Split a strain into segments according to the schedule.
     """
 
-    def __init__(self):
+    def __init__(
+        self,
+        sample_rate: int = None,
+        schedule: torch.Tensor = None,
+        device=None,
+    ):
         super().__init__()
+        self.sample_rate = sample_rate
+        self.schedule = schedule
+        self.device = device
 
-    def build_variable_indices(
-        self, 
-        sr: int, 
-        schedule: torch.Tensor, 
-        device=None
+        if sample_rate is not None and schedule is not None:
+            self._validate_inputs(sample_rate, schedule)
+            self.idx = self.build_variable_indices(
+                sample_rate, schedule, device
+            )
+        else:
+            self.idx = None
+
+    def _validate_inputs(
+        self,
+        sample_rate: int,
+        schedule: torch.Tensor,
+        strain: torch.Tensor = None,
     ):
         """
-        Build a mask of indices to decimate a signal according to the schedule.
+        Validate inputs for compatibility.
+        If `strain` is None, only validates schedule and sample_rate.
         """
+        if not isinstance(sample_rate, int) or sample_rate <= 0:
+            raise ValueError(
+                f"Sample rate must be a positive integer, got {sample_rate}"
+            )
+
+        if schedule.ndim != 2 or schedule.shape[1] != 3:
+            raise ValueError(
+                f"Schedule must be of shape (N, 3), got {schedule.shape}"
+            )
+
+        if not torch.all(schedule[:, 1] > schedule[:, 0]):
+            raise ValueError(
+                "Each schedule segment must have end_time > start_time"
+            )
+
+        if torch.any(schedule[:, 2] <= 0):
+            raise ValueError("Target sample rates must be positive")
+
+        if torch.any(sample_rate % schedule[:, 2].long() != 0):
+            raise ValueError(
+                f"Sample rate {sample_rate} must be divisible by all "
+                f"target rates {schedule[:, 2].tolist()}"
+            )
+
+        if strain is not None:
+            if strain.ndim < 1:
+                raise ValueError(
+                    "strain must have at least 1 dimension (time axis)"
+                )
+
+            expected_len = int(
+                (schedule[:, 1][-1] - schedule[:, 0][0]) * sample_rate
+            )
+            if expected_len != strain.shape[-1]:
+                raise ValueError(
+                    f"Waveform length {strain.shape[-1]} does not match "
+                    f"schedule duration {expected_len}"
+                )
+
+    def build_variable_indices(
+        self,
+        sample_rate: int = None,
[…]
