### 2025-09-26 e692ff3
made changes based on comments and fixed the documentation
