+++ ml4gw/transforms/decimator.py
@@ -2,32 +2,78 @@ import torch
 
 
 class Decimator(torch.nn.Module):
-    """
-    Downsample/decimate timeseries according to a schedule.
+    r"""
+    Downsample (decimate) a timeseries according to a user-defined schedule.
 
-    The schedule is a tensor of shape (N, 3), where each row is:
-    [start_time, end_time, target_sample_rate]
+    .. note::
 
-    Example schedule (seconds, Hz):
-    [[0, 40, 256],   # first 40s at 256 Hz
-     [40, 58, 512],  # next 18s at 512 Hz
-     [58, 60, 2048]] # last 2s at full 2048 Hz
+        This is a naive decimator that does not use any IIR/FIR filtering
+        and selects every M-th sample according to the schedule.
 
-    Strain: Tensor of shape (B, C, T), where
-    B=batch size, C=channels, T=time.
+    The schedule specifies which segments of the input to keep and at what
+    sampling rate. Each row of the schedule has the form:
 
-    Functions:
-        build_variable_indices(sample_rate, schedule, device)
-            Compute the sample indices to keep from the full strain.
-        split_by_schedule(strain, schedule)
-            Split a strain into segments according to the schedule.
+        `[start_time, end_time, target_sample_rate]`
+
+    Args:
+        sample_rate (int):
+            Sampling rate (Hz) of the input timeseries.
+        schedule (torch.Tensor):
+            Tensor of shape `(N, 3)` defining start time, end time,
+            and target sample rate for each segment.
+
+    Shape:
+        - Input: `(B, C, T)` where
+            - B = batch size
+            - C = channels
+            - T = number of timesteps
+                    (must equal schedule duration × sample_rate)
+        - Output:
+            - If ``split=False`` → `(B, C, T')` where `T'` is total
+              number of decimated samples across all segments.
+            - If ``split=True`` → list of tensors, one per segment.
+
+    Returns:
+        torch.Tensor or List[torch.Tensor]:
+            The decimated timeseries, or list of decimated segments if
+            ``split=True``.
+
+    Example:
+        .. code-block:: python
+
+            >>> import torch
+            >>> from ml4gw.transforms.decimator import Decimator
+
+            >>> sample_rate = 2048
+            >>> X_duration = 60
+
+            >>> schedule = torch.tensor(
+            ...     [[0, 40, 256], [40, 58, 512], [58, 60, 2048]],
+            ...     dtype=torch.int,
+            ... )
+
+            >>> decimator = Decimator(sample_rate=sample_rate,
+            ...    schedule=schedule)
+            >>> X = torch.randn(1, 1, sample_rate * X_duration)
+            >>> X_dec = decimator(X)
+            >>> X_seg = decimator(X, split=True)
+
+            >>> print("Original shape:", X.shape)
+            Original shape: torch.Size([1, 1, 122880])
+            >>> print("Decimated shape:", X_dec.shape)
+            Decimated shape: torch.Size([1, 1, 23552])
+            >>> for i, seg in enumerate(X_seg):
+            ...     print(f"Segment {i} shape:", seg.shape)
+            Segment 0 shape: torch.Size([1, 1, 10240])
+            Segment 1 shape: torch.Size([1, 1, 9216])
+            Segment 2 shape: torch.Size([1, 1, 4096])
     """
 
     def __init__(
         self,
         sample_rate: int = None,
         schedule: torch.Tensor = None,
-    ):
+    ) -> None:
         super().__init__()
         self.sample_rate = sample_rate
         self.schedule = schedule
@@ -36,52 +82,38 @@ class Decimator(torch.nn.Module):
         idx = self.build_variable_indices()
[…]
