### 2025-09-28 2d189f2
Add decimator transform for downsampling timeseries data (#226)
* Implements a decimation transform to reduce the sampling rate of input time series based on a schedule
* Add unit test for waveforms against decimation
