### 2025-09-29 7fee071
refactor project metadata
