### 2025-09-29 ef3471f
refactor project metadata (#227)
