### 2025-09-30 f2ca3aa
Add pyupgrade to ruff rules
