### 2025-09-30 caf191c
Add isort to ruff rules
