+++ pyproject.toml
@@ -77,7 +77,7 @@ target-version = "py312"
 [tool.ruff.lint]
 extend-safe-fixes = ["C408"]
 mccabe = {"max-complexity" = 18}
-select = ["B", "C", "E", "F", "UP", "W"]
+select = ["B", "C", "E", "F", "I", "UP", "W"]
 # ignore W605 and E203 for consistency with black
 # ignore F722 for compatibility with jaxtyping
 ignore = ["W605", "E203", "F722"]  
