+++ README.md
@@ -9,40 +9,40 @@
 | ml4gw/constants.py                        |       12 |        0 |    100% |           |
 | ml4gw/dataloading/\_\_init\_\_.py         |        3 |        0 |    100% |           |
 | ml4gw/dataloading/chunked\_dataset.py     |       45 |        2 |     96% |    61, 96 |
-| ml4gw/dataloading/hdf5\_dataset.py        |       73 |        4 |     95% |86, 201-205 |
-| ml4gw/dataloading/in\_memory\_dataset.py  |       63 |        4 |     94% |157-158, 203-204 |
+| ml4gw/dataloading/hdf5\_dataset.py        |       73 |        4 |     95% |86, 199-203 |
+| ml4gw/dataloading/in\_memory\_dataset.py  |       62 |        4 |     94% |155-156, 201-202 |
 | ml4gw/distributions.py                    |      146 |        2 |     99% |   114-115 |
-| ml4gw/gw.py                               |       99 |        0 |    100% |           |
+| ml4gw/gw.py                               |       98 |        0 |    100% |           |
 | ml4gw/nn/\_\_init\_\_.py                  |        0 |        0 |    100% |           |
 | ml4gw/nn/autoencoder/\_\_init\_\_.py      |        3 |        3 |      0% |       1-3 |
-| ml4gw/nn/autoencoder/base.py              |       44 |       44 |      0% |      1-94 |
-| ml4gw/nn/autoencoder/convolutional.py     |       55 |       55 |      0% |     1-159 |
-| ml4gw/nn/autoencoder/skip\_connection.py  |       30 |       30 |      0% |      1-47 |
+| ml4gw/nn/autoencoder/base.py              |       43 |       43 |      0% |      1-90 |
+| ml4gw/nn/autoencoder/convolutional.py     |       54 |       54 |      0% |     1-156 |
+| ml4gw/nn/autoencoder/skip\_connection.py  |       30 |       30 |      0% |      1-45 |
 | ml4gw/nn/autoencoder/utils.py             |       12 |       12 |      0% |      1-15 |
 | ml4gw/nn/norm.py                          |       51 |        2 |     96% |   87, 106 |
 | ml4gw/nn/resnet/\_\_init\_\_.py           |        2 |        0 |    100% |           |
-| ml4gw/nn/resnet/resnet\_1d.py             |      142 |        1 |     99% |       331 |
-| ml4gw/nn/resnet/resnet\_2d.py             |      142 |        1 |     99% |       331 |
+| ml4gw/nn/resnet/resnet\_1d.py             |      143 |        1 |     99% |       330 |
+| ml4gw/nn/resnet/resnet\_2d.py             |      143 |        1 |     99% |       330 |
 | ml4gw/nn/streaming/\_\_init\_\_.py        |        2 |        0 |    100% |           |
-| ml4gw/nn/streaming/online\_average.py     |       49 |        4 |     92% |82, 90-91, 115 |
-| ml4gw/nn/streaming/snapshotter.py         |       35 |        2 |     94% |   96, 110 |
-| ml4gw/spectral.py                         |      147 |        0 |    100% |           |
+| ml4gw/nn/streaming/online\_average.py     |       48 |        4 |     92% |80, 88-89, 113 |
+| ml4gw/nn/streaming/snapshotter.py         |       35 |        2 |     94% |   93, 107 |
+| ml4gw/spectral.py                         |      146 |        0 |    100% |           |
 | ml4gw/transforms/\_\_init\_\_.py          |       10 |        0 |    100% |           |
 | ml4gw/transforms/decimator.py             |       38 |        0 |    100% |           |
-| ml4gw/transforms/iirfilter.py             |       18 |        0 |    100% |           |
+| ml4gw/transforms/iirfilter.py             |       17 |        0 |    100% |           |
[…]
