### 2025-10-01 2a0a018
Fix: adjust HDF5 dataset handling
