### 2025-10-03 #8836 @samaalshammari
sorry for the delay @wbenoit26... I think the changes you made look great. I'm happy to sign off as well except I have one last question. Maybe I wasn't clear but this particular [page](https://ml4gw.github.io/ml4gw/api/ml4gw.html) in the documentation is missing some definitions... I was wondering if you knew about that/ if that is important as part of the review, I guess the second one is for @jatkinson1000 . If that's not an issue I'll tick the final squares and approve everything :D
