### 2025-10-03 #8836 @jatkinson1000
Hey @samaalshammari and @wbenoit26

Thanks for such a thorough review of the package and documentation @samaalshammari
I agree that completing those would be useful.
I _think_ that they're pulled from module docstrings, so if you can add those @wbenoit26 we should then be good to go towards acceptance.
