### 2025-10-03 a3cf9ca
Add module-level docstrings to top-level modules
