+++ ml4gw/waveforms/__init__.py
@@ -1,2 +1,8 @@
+"""
+This module contains tools for generating waveforms,
+both CBC approximants and more generic waveforms like
+sine-gaussians.
+"""
+
 from .adhoc import *
 from .cbc import *
