+++ ml4gw/dataloading/__init__.py
@@ -1,3 +1,8 @@
+"""
+This module contains tools for efficient in-memory and
+out-of-memory dataloading.
+"""
+
 from .chunked_dataset import ChunkedTimeSeriesDataset
 from .hdf5_dataset import Hdf5TimeSeriesDataset
 from .in_memory_dataset import InMemoryDataset
