### 2025-10-03 #8836 @jatkinson1000
@wbenoit26 it looks like the PyTorch reference isn't being rendered correctly in the paper if you can address please.
