+++ README.md
@@ -41,7 +41,7 @@
 | ml4gw/transforms/waveforms.py             |       41 |        0 |    100% |           |
 | ml4gw/transforms/whitening.py             |       47 |        0 |    100% |           |
 | ml4gw/types.py                            |       17 |        0 |    100% |           |
-| ml4gw/utils/interferometer.py             |       20 |        0 |    100% |           |
+| ml4gw/utils/interferometer.py             |       21 |        0 |    100% |           |
 | ml4gw/utils/slicing.py                    |      102 |        0 |    100% |           |
 | ml4gw/waveforms/\_\_init\_\_.py           |        2 |        0 |    100% |           |
 | ml4gw/waveforms/adhoc/\_\_init\_\_.py     |        2 |        0 |    100% |           |
@@ -56,7 +56,7 @@
 | ml4gw/waveforms/cbc/utils.py              |       48 |        0 |    100% |           |
 | ml4gw/waveforms/conversion.py             |       81 |        0 |    100% |           |
 | ml4gw/waveforms/generator.py              |      103 |        0 |    100% |           |
-|                                 **TOTAL** | **3060** |  **227** | **93%** |           |
+|                                 **TOTAL** | **3061** |  **227** | **93%** |           |
 
 
 ## Setup coverage badge
