### 2025-10-03 #8836 @wbenoit26
@jatkinson1000 The released version is v0.7.8 and the DOI is https://doi.org/10.5281/zenodo.17257984. The title/authors/license should all match.
