### 2025-10-04 7c330fd
Update numpy dependency to require version >2.0.0
