+++ pyproject.toml
@@ -28,7 +28,7 @@ dependencies = [
     "jaxtyping>=0.2,<0.3",
     "torch~=2.0",
     "torchaudio~=2.0",
-    "numpy<2.0.0",
+    "numpy>2.0.0",
     "scipy>=1.9.0,<1.15",
 ]
 
