### 2025-10-04 9d00221
Update jaxtyping dependency version constraint
