+++ pyproject.toml
@@ -25,7 +25,7 @@ classifiers = [
 readme = "README.md"
 license = "GPL-3.0-or-later"
 dependencies = [
-    "jaxtyping>=0.2,<0.3",
+    "jaxtyping>=0.2",
     "torch~=2.0",
     "torchaudio~=2.0",
     "numpy>2.0.0",
