### 2025-10-04 7f8c5b0
Update scipy version requirement in pyproject.toml
