+++ pyproject.toml
@@ -29,7 +29,7 @@ dependencies = [
     "torch~=2.0",
     "torchaudio~=2.0",
     "numpy>2.0.0",
-    "scipy>=1.9.0,<1.15",
+    "scipy>=1.16",
 ]
 
 [dependency-groups]
