### 2025-10-12 3c5bfd0
device consistency in UniformInComovingVolume
