+++ ml4gw/distributions.py
@@ -268,8 +268,8 @@ class UniformComovingVolume(dist.Distribution):
         """
         self._generate_distance_grids()
         bounds = torch.tensor(
-            [self.minimum, self.maximum],
-            device=self.minimum.device)
+            [self.minimum, self.maximum], device=self.minimum.device
+        )
         z_min, z_max = self._linear_interp_1d(
             self.distance_grid, self.z_grid, bounds
         )
@@ -280,8 +280,9 @@ class UniformComovingVolume(dist.Distribution):
         """
         Generate distance grids based on the specified redshift range.
         """
-        self.z_grid = torch.linspace(0, self.z_grid_max, self.grid_size,
-                                     device=self.minimum.device)
+        self.z_grid = torch.linspace(
+            0, self.z_grid_max, self.grid_size, device=self.minimum.device
+        )
         self.dz = self.z_grid[1] - self.z_grid[0]
         # C is specfied in m/s, h0 in km/s/Mpc, so divide by 1000 to convert
         comoving_dist_grid = (
@@ -290,8 +291,9 @@ class UniformComovingVolume(dist.Distribution):
             )
             / 1000
         )
-        zero_prefix = torch.zeros(1, dtype=comoving_dist_grid.dtype,
-                                  device=self.minimum.device)
+        zero_prefix = torch.zeros(
+            1, dtype=comoving_dist_grid.dtype, device=self.minimum.device
+        )
         self.comoving_dist_grid = torch.cat([zero_prefix, comoving_dist_grid])
         self.luminosity_dist_grid = self.comoving_dist_grid * (1 + self.z_grid)
 
@@ -321,7 +323,9 @@ class UniformComovingVolume(dist.Distribution):
             p_of_distance, self.distance_grid
         )
         cdf = torch.cumulative_trapezoid(self.pdf, self.distance_grid)
-        zero_prefix = torch.zeros(1, dtype=cdf.dtype, device=self.minimum.device)
+        zero_prefix = torch.zeros(
+            1, dtype=cdf.dtype, device=self.minimum.device
+        )
         self.cdf = torch.cat([zero_prefix, cdf])
         self.log_pdf = torch.log(self.pdf)
 
@@ -348,8 +352,7 @@ class UniformComovingVolume(dist.Distribution):
         )
         inside_range = (value >= self.minimum) & (value <= self.maximum)
         log_prob[~inside_range] = torch.as_tensor(
-            float("-inf"),
-            device=self.minimum.device
+            float("-inf"), device=self.minimum.device
         )
         return log_prob
 
