### 2025-10-14 263624c
device consistency in UniformInComovingVolume (#233)
closes #232
