### 2025-10-21 #8836 @jatkinson1000
@wbenoit26 I've had a final read of the paper - very well written, thank you.
There is one minor typesetting change that I opened the above PR for, then I think you are ready for publication!!
