### 2025-10-21 #8836 @crvernon
@wbenoit26 - I just need you to address the following before I accept this submission for publication:

- [ ] LINE 38:  The "g" in "gaussian" should be capitalized.

Let me know when this change has been made and I'll accept.  Thank you!
