### 2025-10-21 #8836 @crvernon
🥳 Congratulations on your new publication @wbenoit26 ! Many thanks to @jatkinson1000 for editing and @Uddiptaatwork , @samaalshammari , and @Rodrigo-Tenorio for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
